+++ lib/widgets/adventures_in_app.dart
@@ -39,7 +39,7 @@ class AdventuresInApp extends StatelessWidget {
               converter: (store) => store.state.authState,
               builder: (context, authState) {
                 return (authState == AuthState.signedInAndGitHubToken)
-                    ? GitHubSummary()
+                    ? HomePage()
                     : AuthPage();
               },
             ),
